Move the shared boundary mapper to gateway/boundary.py
service.py crossed the 999-line repo limit with the mapper inline; the
shared home also spares the bridge a service import.
